requirements.txt
# Core Web Framework
Flask
Flask-Login
Flask-SQLAlchemy

# NLP and AI
spacy
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl

# Document Processing
PyPDF2
python-docx
PyMuPDF

# Utilities
requests
python-dotenv
greenlet
